+++ nonconform/__init__.py
@@ -42,7 +42,7 @@ Examples:
     ... )
 """
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 __author__ = "Oliver Hennhoefer"
 __email__ = "oliver.hennhoefer@mail.de"